ML/picturePredict/requirements.txt
tensorflow
numpy
opencv-python
streamlit
pillow